restructure: components
